Allow `jedi_hover` and `jedi_signature_help` plugins to hide docstrings.
Both the plugins include a signature's docstring by the default in the
content they generate. Which is cool, as long as the docstring's author
doesn't moonlight as a novelist. Otherwise, the content rendered in the
popup will cover the entire screen. It's annoying to say the least.

This changeset allows a user to hide the docstrings by adding config
flags. They default to `true` so people aren't surprised by the change.
I've been running this version in ST4 with no issues for a while, so I'm
submitting it as a PR.